refactor(*): `로또앱`과 앱 외부의 `공통`으로 패키지를 정리
