+++ src/main/java/lotto/lottoapp/value/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.value;
+package lotto.lottoapp.value;
 
 import java.util.Objects;
