+++ src/main/java/lotto/lottoapp/value/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lotto.value;
+package lotto.lottoapp.value;
 
 import java.util.List;
 import java.util.stream.Stream;
