+++ src/test/java/lotto/LottoTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.model.Lotto;
-import lotto.value.LottoNumber;
+import lotto.lottoapp.model.Lotto;
+import lotto.lottoapp.value.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
