+++ src/test/java/lotto/value/LottoNumberTest.java
@@ -2,6 +2,7 @@ package lotto.value;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import lotto.lottoapp.value.LottoNumber;
 import org.junit.jupiter.api.Test;
 
 class LottoNumberTest {
